+++ profileservice/restservice/profile_rest_service.py
@@ -1,13 +1,19 @@
+import sys
+sys.path.append('../../')
+
+import datetime
 import logging
 import flask
-import uuid
+import uuid as uuidlib
 
-import profileservice.restservice.utils.mongoutils as mongoutils
 import profileservice.configs as cfg
+import profileservice.restservice.utils.mongoutils as mongoutils
 import profileservice.restservice.utils.jsonutils as jsonutils
+import profileservice.restservice.utils.datasetutils as datasetutils
 
 from bson import ObjectId
 from flask import flash, redirect, jsonify, make_response, request
+
 from profileservice.dao.pii_data import pii_data
 from profileservice.dao.non_pii_data import non_pii_data
 from profileservice.restservice.utils.otherutils import create_file_descriptor
@@ -20,39 +26,44 @@ __logger = logging.getLogger("profileservice")
 """
 rest service for root directory
 """
-@app.route('/profiles', methods=['GET', 'POST'])
+@app.route('/profiles', methods=['POST'])
 def non_pii_root_dir():
-    if request.method == 'GET':
-        db_data = mongoutils.db.non_pii_collection.find({}, {'_id': False})
-        data_list = list(db_data)
+    # if request.method == 'GET':
+    #     db_data = mongoutils.db.non_pii_collection.find({}, {cfg.FIELD_OBJECTID: False})
+    #     data_list = list(db_data)
+    #
+    #     out_json = mongoutils.construct_json_from_query_list(data_list)
+    #     logging.debug("list all profiles")
+    #
+    #     return out_json
 
-        out_json = mongoutils.construct_json_from_query_list(data_list)
-        logging.debug("list all profiles")
-
-        return out_json
-
-    elif request.method == 'POST':
-        is_new_install = False
-
-        try:
-            in_json = request.get_json()
-        except:
-            is_new_install = True
+    if request.method == 'POST':
+        is_new_install = True
 
         # check if uuid is in there otherwise it is either a first installation
         try:
+            in_json = request.get_json()
             non_pii_uuid = in_json["uuid"]
             # even if there is non_pii_uuid is in the input json, it could be a new one
             # check if the dataset is existing with given uuid
-            dataset = mongoutils.get_non_pii_dataset_from_field('uuid', non_pii_uuid)
-            if dataset is None:
-                is_new_install = True
+            dataset = mongoutils.get_non_pii_dataset_from_field(cfg.FIELD_PROFILE_UUID, non_pii_uuid)
+            if dataset is not None:
+                is_new_install = False
+                msg = "UUID in input json already exists in the database " + str(non_pii_uuid)
+                logging.error(msg)
+                return bad_request()
         except:
-            is_new_install = True
+            pass
 
         if is_new_install:
             # new installation of the app
+            currenttime = datetime.datetime.now()
+            currenttime = currenttime.strftime("%Y/%m/%dT%H:%M:%S")
             non_pii_dataset = non_pii_data('')
+            non_pii_uuid = str(uuidlib.uuid4())
+            non_pii_dataset.set_uuid(non_pii_uuid)
+            non_pii_dataset.set_first_modified(currenttime)
+            non_pii_dataset.set_last_modified(currenttime)
             dataset, id = mongoutils.insert_non_pii_dataset_to_mongodb(non_pii_dataset)
             profile_uuid = dataset["uuid"]
             dataset = jsonutils.remove_objectid_from_dataset(dataset)
@@ -60,66 +71,12 @@ def non_pii_root_dir():
             msg = "new profile with new uuid has been created: " + str(profile_uuid)
             logging.debug(msg)
 
-            return out_json
-        else:
-            # updated non-pii profile data with given uuid
-            if len(non_pii_uuid) > 0:
-                # get dataset using uuid
-                dataset = mongoutils.get_non_pii_dataset_from_field('uuid', non_pii_uuid)
-                if dataset is None:
-                    msg = "the dataset does not exist with uuid of : " + str(non_pii_uuid)
-                    logging.error(msg)
-                    return not_found()
-                else:
-                    # TODO check the app that the app should post cached data fully to the endpoint
-                    non_pii_dataset = non_pii_data(in_json)
-
-                    # update non_pii object
-                    result, dataset = mongoutils.update_non_pii_dataset_in_mongo_by_field('uuid', non_pii_uuid, non_pii_dataset)
-
-                    if (result):
-                        out_json = mongoutils.construct_json_from_query_list(dataset)
-                        msg = "New profile has been posted with : " + str(non_pii_uuid)
-                        logging.debug(msg)
-
-                        return out_json
-                    else:
-                        return bad_request()
-            else:
-                msg = "uuid " + str(non_pii_uuid) + " not found"
-                logging.error(msg)
-                return not_found()
-
-# """
-# post a new record to get non-pii (when new app get installed
-# """
-# @app.route('/profiles/non-pii', methods=['POST'])
-# def new_app_installation():
-#     if request.method == 'POST':
-#         non_pii_dataset = non_pii_data('')
-#         dataset, id = mongoutils.insert_non_pii_dataset_to_mongodb(non_pii_dataset)
-#         uuid = dataset["uuid"]
-#         outstr = "{\"uuid\": \"%s\"}" % (uuid)
-#         outjson = json.loads(outstr)
-#         data_dump = dumps(outjson)
-#         out_json = make_response(data_dump)
-#         out_json.mimetype = 'application/json'
-#
-#         msg = "new profile with new uuid has been created: " + str(uuid)
-#         logging.debug(msg)
-#
-#         return out_json
-#
-#     else:
-#         bad_request()
-#
-#         return None
-
+            return return_id('uuid', profile_uuid)
 
 """
 provide profile information by profile id or remove it
 """
-@app.route('/profiles/<uuid>', methods=['GET', 'DELETE'])
+@app.route('/profiles/<uuid>', methods=['GET', 'DELETE', 'PUT'])
 def deal_profile_id(uuid):
     if uuid != None:
         is_objectid = mongoutils.check_if_objectid(uuid)
@@ -129,7 +86,7 @@ def deal_profile_id(uuid):
             id = ObjectId(uuid)
             db_data = mongoutils.query_non_pii_dataset_by_objectid(id)
         else:
-            db_data = mongoutils.query_non_pii_dataset('uuid', uuid)
+            db_data = mongoutils.query_non_pii_dataset(cfg.FIELD_PROFILE_UUID, uuid)
 
         data_list = list(db_data)
         if len(data_list) > 0:
@@ -139,19 +96,56 @@ def deal_profile_id(uuid):
                 msg = "request profile information: " + str(uuid)
                 logging.debug(msg)
 
+            if request.method == 'PUT':
+                try:
+                    in_json = request.get_json()
+                except:
+                    return bad_request()
+
+                # check if the uuid is really existing in the database
+                non_pii_dataset = mongoutils.get_non_pii_dataset_from_field(cfg.FIELD_PROFILE_UUID, uuid)
+
+                if non_pii_dataset is None:
+                    msg = "There is no profile dataset with given uuid " + str(uuid)
+                    logging.error(msg)
+                    return not_found()
+                else:
+                    msg = "Profile data will be updated with the id of " + str(uuid)
+                    logging.debug(msg)
+
+                    non_pii_dataset = datasetutils.update_non_pii_dataset_from_json(non_pii_dataset, in_json)
+                    currenttime = datetime.datetime.now()
+                    currenttime = currenttime.strftime("%Y/%m/%dT%H:%M:%S")
+                    non_pii_dataset.set_last_modified(currenttime)
+
+                    result, non_pii_dataset = mongoutils.update_non_pii_dataset_in_mongo_by_field(cfg.FIELD_PROFILE_UUID, uuid,
+                                                                                              non_pii_dataset)
+
+                    if result is None:
+                        msg = "Failed to update non Profile dataset: " + str(uuid)
+                        logging.error(msg)
+
+                        return not_implemented()
+                    else:
+                        out_json = mongoutils.construct_json_from_query_list(non_pii_dataset)
+                        msg = "Profile data has been posted with : " + str(uuid)
+                        logging.debug(msg)
+
+                        return out_json
+
             # delete profile by using profile id
             if request.method == 'DELETE':
                 if (is_objectid):
-                    mongoutils.db.non_pii_collection.delete_one({'_id': id})
-                    msg = "deleted profile information: " + str(uuid)
+                    mongoutils.db.non_pii_collection.delete_one({cfg.FIELD_OBJECTID: id})
+                    msg = "deleted profile information: " + str(id)
                     logging.debug(msg)
-                    return entry_deleted()
+                    return entry_deleted(id)
                 else:
                     try:
-                        mongoutils.db.non_pii_collection.delete_one({'uuid': uuid})
+                        mongoutils.db.non_pii_collection.delete_one({cfg.FIELD_PROFILE_UUID: uuid})
                         msg = "deleted profile information: " + str(uuid)
                         logging.debug(msg)
-                        return entry_deleted()
+                        return entry_deleted(uuid)
                     except:
                         msg = "failed to deleted. not found: " + str(uuid)
                         logging.error(msg)
@@ -172,7 +166,7 @@ upload image for the profile
 def upload_profile_image(uuid):
     # TODO add unsupported media type handler
     if request.method == 'POST':
-        non_pii_dataset = mongoutils.get_non_pii_dataset_from_field('uuid', uuid)
+        non_pii_dataset = mongoutils.get_non_pii_dataset_from_field(cfg.FIELD_PROFILE_UUID, uuid)
         if non_pii_dataset is None:
             msg = "the dataset does not exist: " + str(uuid)
             logging.error(msg)
@@ -196,8 +190,11 @@ def upload_profile_image(uuid):
             file_descriptors.append(fd)
             non_pii_dataset.set_file_descriptors(file_descriptors)
             non_pii_dataset.set_image_uri(fd.dataURL)
+            currenttime = datetime.datetime.now()
+            currenttime = currenttime.strftime("%Y/%m/%dT%H:%M:%S")
+            non_pii_dataset.set_last_modified(currenttime)
 
-            result, non_pii_dataset = mongoutils.update_non_pii_dataset_in_mongo_by_field('uuid', uuid, non_pii_dataset)
+            result, non_pii_dataset = mongoutils.update_non_pii_dataset_in_mongo_by_field(cfg.FIELD_PROFILE_UUID, uuid, non_pii_dataset)
 
             if (result):
                 out_json = mongoutils.construct_json_from_query_list(non_pii_dataset)
@@ -223,7 +220,7 @@ def pii_root_dir():
 
         # TODO this if else method should smarter like case or something else
         if term_uuid != None:
-            out_json = mongoutils.get_pii_http_output_query_result_using_field_string('pii_uuid', term_uuid)
+            out_json = mongoutils.get_pii_http_output_query_result_using_field_string(cfg.FIELD_PID, term_uuid)
             if out_json == None:
                 return not_found()
             else:
@@ -247,102 +244,89 @@ def pii_root_dir():
             else:
                 return out_json
 
-        else:
-            db_data = mongoutils.db.pii_collection.find({}, {'_id': False})
-            data_list = list(db_data)
+        return bad_request()
 
         out_json = mongoutils.construct_json_from_query_list(data_list)
         logging.debug("list all pii data")
 
         return out_json
+
     elif request.method == 'POST':
         is_new_entry = False
         try:
             in_json = request.get_json()
         except:
-            #TODO if there is no input json, this should throw an error
-            #is_new_entry = True
             bad_request()
-
 
         # get uuid, if failed it is a bad request
         try:
-            non_pii_uuid = in_json["uuid"]
+            non_pii_uuid = in_json[cfg.FIELD_PROFILE_UUID]
         except:
             bad_request()
 
         # check if it is a new record or existing record
         try:
-            pii_uuid = in_json["pii_uuid"]
+            pid = in_json[cfg.FIELD_PID]
+            dataset = mongoutils.get_pii_dataset_from_field(cfg.FIELD_PROFILE_UUID, pid)
         except:
+            dataset = None
             is_new_entry = True
 
+        # check if the email already exists
+        try:
+            email = in_json['email']
+            dataset = mongoutils.get_pii_dataset_from_field('email', email)
+            if dataset is not None:
+                pid = dataset.get_pid()
+                return return_id('pid', pid)
+        except:
+            pass
+
+        # check if the phonenumber already exists
+        try:
+            phone = in_json['phone']
+            dataset = mongoutils.get_pii_dataset_from_field('phone', phone)
+            if dataset is not None:
+                pid = dataset.get_pid()
+                return return_id('pid', pid)
+        except:
+            pass
+
+        if dataset is not None:
+            is_new_entry = False
+
         pii_dataset = pii_data(in_json)
+
         if is_new_entry:
             # insert new pii_dataset
-            pii_uuid = str(uuid.uuid4())
-            pii_dataset.set_pii_uuid(pii_uuid)
+            currenttime = datetime.datetime.now()
+            currenttime = currenttime.strftime("%Y/%m/%dT%H:%M:%S")
+            pid = str(uuidlib.uuid4())
+            pii_dataset.set_pid(pid)
             non_pii_uuid_from_dataset = []
             non_pii_uuid_from_dataset.append(non_pii_uuid)
-            pii_dataset.set_non_pii_uuid(non_pii_uuid_from_dataset)
+            pii_dataset.set_uuid(non_pii_uuid_from_dataset)
+            pii_dataset.set_first_modified(currenttime)
+            pii_dataset.set_last_modified(currenttime)
             pii_dataset = mongoutils.insert_pii_dataset_to_mongodb(pii_dataset)
 
             if pii_dataset is None:
-                msg = "Failed to update non pii uuid into pii dataset: " + str(pii_uuid)
+                msg = "Failed to update non pii uuid into pii dataset: " + str(pid)
                 logging.error(msg)
 
                 return not_implemented()
             else:
                 pii_dataset = jsonutils.remove_objectid_from_dataset(pii_dataset)
                 out_json = mongoutils.construct_json_from_query_list(pii_dataset)
-                msg = "Pii data has been posted with : " + str(pii_uuid)
+                msg = "Pii data has been posted with : " + str(pid)
                 logging.debug(msg)
 
-                return out_json
+                return return_id('pid', pid)
         else:
-            # TODO check the app that the app should post cached data fully to the endpoint
-            # if pii_uuid is none or the lenght is zero, then it is not right
-            if pii_uuid is None or len(pii_uuid) == 0:
-                return not_found()
-            # check if the pii_uuid is really existing in the database
-            pii_dataset = mongoutils.get_pii_dataset_from_field('pii_uuid', pii_uuid)
+            msg = 'The request is wrong or the entry already exists'
+            logging.error(msg)
 
-            if pii_dataset == None:
-                return not_found()
-            else:
-                msg = "Pii data will be updated with the id of " + str(pii_uuid)
-                logging.debug(msg)
-
-                # update pii_dataset's non_pii_uuid
-                non_pii_uuid_from_dataset = pii_dataset.get_non_pii_uuid()
-                is_non_pii_uuid_in_json_new = True
-                if non_pii_uuid_from_dataset is None:
-                    non_pii_uuid_from_dataset = []
-                else:   # check if non-pii-uuid is already in there
-                    for i in range(len(non_pii_uuid_from_dataset)):
-                        if non_pii_uuid == non_pii_uuid_from_dataset[i]:
-                            is_non_pii_uuid_in_json_new = False
-
-                # adde non-pii uuid in json only if it is now uuid
-                if is_non_pii_uuid_in_json_new:
-                    non_pii_uuid_from_dataset.append(non_pii_uuid)
-
-                pii_dataset.set_non_pii_uuid(non_pii_uuid)
-
-                result, pii_dataset = mongoutils.update_pii_dataset_in_mongo_by_field('pii_uuid', pii_uuid,
-                                                                                      pii_dataset)
-
-                if result is None:
-                    msg = "Failed to update non pii uuid into pii dataset: " + str(pii_uuid)
-                    logging.error(msg)
-
-                    return not_implemented()
-                else:
-                    out_json = mongoutils.construct_json_from_query_list(pii_dataset)
-                    msg = "Pii data has been posted with : " + str(pii_uuid)
-                    logging.debug(msg)
-
-                    return out_json
+            return bad_request()
 
     else:
         logging.error("list pii dataset failed.")
@@ -351,49 +335,104 @@ def pii_root_dir():
 """
 provide profile information by profile id or remove it
 """
-@app.route('/profiles/pii/<uuid>', methods=['GET', 'DELETE'])
-def deal_pii_id(uuid):
-    if uuid != None:
-        is_objectid = mongoutils.check_if_objectid(uuid)
+@app.route('/profiles/pii/<pid>', methods=['GET', 'DELETE', 'PUT'])
+def deal_pid(pid):
+    if pid != None:
+        is_objectid = mongoutils.check_if_objectid(pid)
 
         # query using either non-pii ObjectId or uuid
         if (is_objectid):
-            id = ObjectId(uuid)
+            id = ObjectId(pid)
             db_data = mongoutils.query_pii_dataset_by_objectid(id)
         else:
-            db_data = mongoutils.query_pii_dataset('pii_uuid', uuid)
+            db_data = mongoutils.query_pii_dataset(cfg.FIELD_PID, pid)
 
         data_list = list(db_data)
         if len(data_list) > 0:
             out_json = mongoutils.construct_json_from_query_list(data_list)
 
             if request.method == 'GET':
-                msg = "request profile information: " + str(uuid)
+                msg = "request profile information: " + str(pid)
                 logging.debug(msg)
+
+            # update the pii information
+            if request.method == 'PUT':
+                try:
+                    in_json = request.get_json()
+                except:
+                    bad_request()
+
+                # check if the pid is really existing in the database
+                pii_dataset = mongoutils.get_pii_dataset_from_field(cfg.FIELD_PID, pid)
+
+                if pii_dataset == None:
+                    msg = "There is no dataset with given pii uuid " + str(pid)
+                    logging.error(msg)
+                    return not_found()
+                else:
+                    msg = "Pii data will be updated with the id of " + str(pid)
+                    logging.debug(msg)
+
+                    pii_dataset = datasetutils.update_pii_dataset_from_json(pii_dataset, in_json)
+                    currenttime = datetime.datetime.now()
+                    currenttime = currenttime.strftime("%Y/%m/%dT%H:%M:%S")
+                    pii_dataset.set_last_modified(currenttime)
+
+                    # update pii_dataset's non_pii_uuid
+                    non_pii_uuid_from_dataset = pii_dataset.get_uuid()
+                    try:
+                        non_pii_uuid = in_json[cfg.FIELD_PROFILE_UUID]
+                        is_non_pii_uuid_in_json_new = True
+                        # check if non-pii-uuid is already in there
+                        for i in range(len(non_pii_uuid_from_dataset)):
+                            if non_pii_uuid == non_pii_uuid_from_dataset[i]:
+                                is_non_pii_uuid_in_json_new = False
+
+                        # adde non-pii uuid in json only if it is now uuid
+                        if is_non_pii_uuid_in_json_new:
+                            non_pii_uuid_from_dataset.append(non_pii_uuid)
+                            pii_dataset.set_non_pii_uuid(non_pii_uuid)
+                    except:
+                        pass
+
+                    result, pii_dataset = mongoutils.update_pii_dataset_in_mongo_by_field(cfg.FIELD_PID, pid,
+                                                                                          pii_dataset)
+
+                    if result is None:
+                        msg = "Failed to update non pii uuid into pii dataset: " + str(pid)
+                        logging.error(msg)
+
+                        return not_implemented()
+                    else:
+                        out_json = mongoutils.construct_json_from_query_list(pii_dataset)
+                        msg = "Pii data has been posted with : " + str(pid)
+                        logging.debug(msg)
+
+                        return out_json
 
             # delete profile by using profile id
             if request.method == 'DELETE':
                 if (is_objectid):
-                    mongoutils.db.pii_collection.delete_one({'_id': id})
-                    msg = "deleted pii information: " + str(uuid)
+                    mongoutils.db_pii.pii_collection.delete_one({cfg.FIELD_OBJECTID: id})
+                    msg = "deleted pii information: " + str(pid)
                     logging.debug(msg)
 
-                    return entry_deleted()
+                    return entry_deleted(id)
                 else:
                     try:
-                        mongoutils.db.pii_collection.delete_one({'pii_uuid': uuid})
-                        msg = "deleted pii information: " + str(uuid)
+                        mongoutils.db_pii.pii_collection.delete_one({cfg.FIELD_PID: pid})
+                        msg = "deleted pii information: " + str(pid)
                         logging.debug(msg)
 
-                        return entry_deleted()
+                        return entry_deleted(pid)
                     except:
-                        msg = "failed to deleted pii. not found: " + str(uuid)
+                        msg = "failed to deleted pii. not found: " + str(pid)
                         logging.error(msg)
                     return not_found()
 
             return out_json
         else:
-            msg = "the pii dataset does not exist: " + str(uuid)
+            msg = "the pii dataset does not exist: " + str(pid)
             logging.error(msg)
             return not_found()
     else:
@@ -402,6 +441,16 @@ def deal_pii_id(uuid):
 """
 make reponse for handling 202 entry deleted
 """
+def return_id(msg, id):
+    message = {
+        'status': 200,
+        'message': msg + ': ' + id,
+    }
+    resp = jsonify(message)
+    resp.status_code = 200
+
+    return make_response(resp)
+
 def entry_deleted(id):
     message = {
         'status': 202,
@@ -484,4 +533,4 @@ def not_implemented(error=None):
 
 
 if __name__ == '__main__':
-    app.run(host='0.0.0.0', debug=True)
+    app.run(host='0.0.0.0', port=5000, debug=True)